+++ entrypoint.sh
@@ -1,5 +1,6 @@
 #!/usr/bin/env bash
 #set -x
+# ([^:]*)$   - from arn
 
 root_folder="$HOME/.aws"
 mkdir -p ${root_folder}
@@ -26,7 +27,7 @@ CHECKSIZE=1023
 # Set internal variable
 declare -i statuscode
 declare -i trycounter
-datestring=$(date +%Y%m%d)
+datestring=$(date +%Y-%m-%d)
 datetimestring=$(date +%Y%m%d%H%M)
 
 function info () {
@@ -65,10 +66,10 @@ function getRdsArn () {
     --resource-type-filters rds:db \
     --query 'ResourceTagMappingList[*].[ResourceARN]' \
     --tag-filters Key=Anemometer,Values=true \
-    --output text
+    --output text | grep -Po '([^:]*)$'
 }
 
-function describeRDS () {
+function describeEngine () {
     local profile=$1
     local region=$2
     local rdsarn=$3
@@ -76,97 +77,130 @@ function describeRDS () {
     /usr/bin/aws rds describe-db-instances \
     --profile=${profileID} \
     --region ${REGION} \
-    --filters "Name=db-instance-id,Values=${rdsArn}" \
-    --query 'DBInstances[*].[DBInstanceIdentifier]' \
+    --db-instance-identifier ${rdsName} \
+    --query 'DBInstances[*].[Engine]' \
     --output text
-
 }
 
 for hour in $(seq 0 23) ; do suffixes="${suffixes} log.$hour" ; done
+for hour in $(seq -f "%02g" 0 23) ; do suffixesdate="${suffixesdate} log.${datestring}.$hour" ; done
+
 info "INFO: ${ENV_NAME} allDB The list of suffixes for slowlogs files ${suffixes} were created"
+info "INFO: ${ENV_NAME} allDB The list of suffixes for Aurora RDS slowlogs files ${suffixesdate} were created"
 info "INFO: ${ENV_NAME} allDB Lets find RDS ARNs  where tag Anemometer=true"
 
 profileIDs=$( grep -Po '(?<=\[)[^]]+(?=\])' ${path_credentials})
 echo "${profileIDs[@]}"
-for profileID in ${profileIDs[@]}; do
+for profileID in ${profileIDs[@]} ; do
     ENV_NAME=${profileID}
     info "INFO: ${ENV_NAME} allDB Lets start work in ${profileID} "
-    rdsArns=$(getRdsArn ${profileID} ${REGION} )
-    info "INFO: ${ENV_NAME} allDB Such arns was found : ${rdsArns}"
-    for rdsArn in ${rdsArns}; do
-        info "INFO: ${ENV_NAME} allDB Lets describe RDS instance name in the account for ${rdsArn}"
-        describedRDS=$(describeRDS ${profileID} ${REGION} ${rdsArn} )
-        info "INFO: ${ENV_NAME} allDB such instance was found : ${describedRDS}"
-        for instanceID in ${describedRDS}; do
-           nextstep="yes"
-           commontemporary="/tmp/slow-${instanceID}-$datestring.log"
-           info "INFO: ${ENV_NAME} ${instanceID} Start to download slowlogs"
-            for suff in ${suffixes} ; do
-                temporaryfile="/tmp/slow-${instanceID}.${suff}"
+    rdsNames=$( getRdsArn ${profileID} ${REGION})
+    info "INFO: ${ENV_NAME} allDB Such RDS with tag Anemometer was found : ${rdsNames}"
+    for rdsName in ${rdsNames}; do
+        info "INFO: ${ENV_NAME} allDB Lets check RDS engine for ${rdsName}"
+        engineRDS=$(describeEngine ${profileID} ${REGION} ${rdsName} )
+        info "INFO: ${ENV_NAME} ${rdsName} The rds ${rdsName} engine type is ${engineRDS} "
+        commontemporary="/tmp/generalslow-${rdsName}-$datestring.log"
+        if [ "${engineRDS}" == "aurora-mysql" ] ; then
+            info "INFO: ${ENV_NAME} ${rdsName} Lets start to download slowlogs for Aurora RDS ${rdsName}, engine is ${engineRDS} "
+            nextstep="yes"
+            for suffdate in ${suffixesdate} ; do
+                temporaryfile="/tmp/slow-${rdsName}.${suffdate}"
                 trycounter=0
                 while [ ${trycounter} -lt 5 ] ; do
-                    info "INFO: ${ENV_NAME} ${instanceID} Downloading  ${temporaryfile}"
-                    info "INFO: ${ENV_NAME} ${instanceID} Free disk space before downloading df -h " : $(df -h)
-                    downloadLogs=$(downloadLog ${profileID} ${REGION} ${instanceID} slowquery/mysql-slowquery."${suff}" ${temporaryfile})
+                    info "INFO: ${ENV_NAME} ${rdsName} Downloading  ${temporaryfile}"
+                    info "INFO: ${ENV_NAME} ${rdsName} Free disk space before downloading df -h " : $(df -h)
+                    downloadLogs=$(downloadLog ${profileID} ${REGION} ${rdsName} slowquery/mysql-slowquery."${suffdate}" ${temporaryfile})
                     statuscode=$?
                     sleep 2
-                    echo "INFO: ${REGION} ${instanceID} ${temporaryfile} stat information " : $(stat -f ${temporaryfile})
-                    info "INFO: ${ENV_NAME} ${instanceID} downloadLogs function statuscode=${statuscode}"
+                    info "INFO: ${ENV_NAME} ${rdsName} downloadLogs function statuscode=${statuscode}"
                     if [ ${statuscode} -gt 0 ] ; then
-                        echo "INFO: ${ENV_NAME} ${instanceID} stdout log file"
-                        echo "CRITICAL: ${ENV_NAME} ${instanceID} An error occurred (DBLogFileNotFoundFault) when calling the DownloadDBLogFilePortion operation: DBLog File: slow-log file is not found on the ${instanceID}. The problem file is slowquery/mysql-slowquery."${suff}" "
+                        echo "INFO: ${ENV_NAME} ${rdsName} stdout log file"
+                        echo "CRITICAL: ${ENV_NAME} ${rdsName} An error occurred (DBLogFileNotFoundFault) when calling the DownloadDBLogFilePortion operation: DBLog File: slow-log file is not found on the ${instanceID}. The problem file is slowquery/mysql-slowquery."${suff}" "
                         ((trycounter++))
-                        info "INFO: ${ENV_NAME} ${instanceID} counter=${trycounter}"
+                        info "INFO: ${ENV_NAME} ${rdsName} counter=${trycounter}"
                         sleep 2
                     else
-                        workfolder="${ENV_NAME}/${instanceID}/${datestring}"
-                        workfolderTime="${ENV_NAME}/${instanceID}/${datetimestring}"
-                        checkfolderonS3=$( aws s3 ls --profile=${ENV_ORIGIN_NAME} s3://${S3_BUCKET}/slowlogs/${workfolder}/ )
-                        statuscode=$?
-                        if [ ${statuscode} -gt 0 ] ; then
-                            info "INFO: ${ENV_NAME} ${instanceID} The folder ${workfolder} on S3 ${S3_BUCKET} is already exist. Will copy to the new folder ${workfolderTime} "
-                            cpslowToS3=$( aws s3 cp ${temporaryfile}  s3://${S3_BUCKET}/slowlogs/${workfolderTime}/ )
-                        else
-                            info "INFO: ${ENV_NAME} ${instanceID} Let's copy ${temporaryfile} to S3 bucket into ${workfolder}"
-                            cpslowToS3=$( aws s3 cp ${temporaryfile}  s3://${S3_BUCKET}/slowlogs/${workfolder}/ )
-                        fi
+                        workfolder="${ENV_NAME}/${rdsName}/${datestring}"
+                        cpslowToS3=$( aws s3 cp --profile=${ENV_ORIGIN_NAME} ${temporaryfile}  s3://${S3_BUCKET}/slowlogs/${workfolder}/ )
                         temporaryfilesize=$(stat -c%s "$temporaryfile")
                         if [[ ${temporaryfilesize} -le ${CHECKSIZE} ]] ; then
-                            echo "ERROR: ${ENV_NAME} ${instanceID} The problem is with downloading ${temporaryfile}. The files size is less than ${CHECKSIZE} bytes"
+                            echo "ERROR: ${ENV_NAME} ${rdsName} The problem is with downloading ${temporaryfile}. The files size is less than ${CHECKSIZE} bytes"
                             ((trycounter++))
                             sleep 2
                         else
-                            info "INFO: ${ENV_NAME} ${instanceID} Downloading finished OK. The size of ${temporaryfile} = ${temporaryfilesize} bytes. Start to add it into  ${commontemporary}"
-                            info "INFO: ${ENV_NAME} ${instanceID} Free disk space after downloading ${temporaryfile} df -h " : $(df -h)
+                            info "INFO: ${ENV_NAME} ${rdsName} Downloading finished OK. The size of ${temporaryfile} = ${temporaryfilesize} bytes. Start to add it into  ${commontemporary}"
+                            info "INFO: ${ENV_NAME} ${rdsName} Free disk space after downloading ${temporaryfile} df -h " : $(df -h)
                             cat ${temporaryfile} >> ${commontemporary}
                             echo >> ${commontemporary}
                             rm -r ${temporaryfile}
                             commontemporaryfilesize=$(stat -c%s "$commontemporary")
-                            info "INFO: ${ENV_NAME} ${instanceID} Size of collecting file is $commontemporary = $commontemporaryfilesize bytes."
+                            info "INFO: ${ENV_NAME} ${rdsName} Size of collecting file is $commontemporary = $commontemporaryfilesize bytes."
                             trycounter=10
                         fi
                     fi
                 done
             done
-            info "INFO: ${ENV_NAME} ${instanceID} Finished downloading ${instanceID} slowlogs by hours"
-            if [[ "${nextstep}" = "yes" && -f "${commontemporary}" ]] ; then
-                info "INFO: ${ENV_NAME} ${instanceID}. Starting to digest collected file ${commontemporary} and add result into anemometer database"
-                /usr/bin/pt-query-digest --user=$ANEMOMETER_MYSQL_USER --password=$ANEMOMETER_MYSQL_PASSWORD \
-                                        --review h=$ANEMOMETER_MYSQL_HOST,D=$ANEMOMETER_MYSQL_DB,t=global_query_review \
-                                        --history h=$ANEMOMETER_MYSQL_HOST,D=$ANEMOMETER_MYSQL_DB,t=global_query_review_history \
-                                        --no-report --limit=0% \
-                                        --filter=" \$event->{Bytes} = length(\$event->{arg}) and \$event->{hostname}=\"${instanceID}\"" \
-                                        ${commontemporary}
-                statuscode=$?
-                info "statuscode=${statuscode} of percona digest tool"
-                if [ ${statuscode} -gt 0 ] ; then
-                    echo "CRITICAL: ${ENV_NAME} ${instanceID} slowlogs digest problem. the statuscode=${statuscode} "
-                else
-                    info "INFO: ${ENV_NAME} ${instanceID} Digest of ${commontemporary} was successful"
-                    rm -f "${commontemporary}"
-                fi
+        else   # aurora finish
+            nextstep="yes"
+            info "INFO: ${ENV_NAME} ${rdsName} Lets start to download slowlogs for ${rdsName}"
+            for suff in ${suffixes} ; do
+                temporaryfile="/tmp/slow-${rdsName}.${suff}"
+                trycounter=0
+                while [ ${trycounter} -lt 5 ] ; do
+                    info "INFO: ${ENV_NAME} ${rdsName} Downloading  ${temporaryfile}"
+                    info "INFO: ${ENV_NAME} ${rdsName} Free disk space before downloading df -h " : $(df -h)
+                    downloadLogs=$(downloadLog ${profileID} ${REGION} ${rdsName} slowquery/mysql-slowquery."${suff}" ${temporaryfile})
+                    statuscode=$?
+                    sleep 2
+                    echo "INFO: ${REGION} ${rdsName} ${temporaryfile} stat information " : $(stat -f ${temporaryfile})
+                    info "INFO: ${ENV_NAME} ${rdsName} downloadLogs function statuscode=${statuscode}"
+                    if [ ${statuscode} -gt 0 ] ; then
+                        echo "INFO: ${ENV_NAME} ${rdsName} stdout log file"
+                        echo "CRITICAL: ${ENV_NAME} ${rdsName} An error occurred (DBLogFileNotFoundFault) when calling the DownloadDBLogFilePortion operation: DBLog File: slow-log file is not found on the ${instanceID}. The problem file is slowquery/mysql-slowquery."${suff}" "
+                        ((trycounter++))
+                        info "INFO: ${ENV_NAME} ${rdsName} counter=${trycounter}"
+                        sleep 2
+                    else
+                        workfolder="${ENV_NAME}/${rdsName}/${datestring}"
+                        cpslowToS3=$( aws s3 cp --profile=${ENV_ORIGIN_NAME} ${temporaryfile}  s3://${S3_BUCKET}/slowlogs/${workfolder}/ )
+                        temporaryfilesize=$(stat -c%s "$temporaryfile")
+                        if [[ ${temporaryfilesize} -le ${CHECKSIZE} ]] ; then
+                            echo "ERROR: ${ENV_NAME} ${rdsName} The problem is with downloading ${temporaryfile}. The files size is less than ${CHECKSIZE} bytes"
+                            ((trycounter++))
+                            sleep 2
+                        else
+                            info "INFO: ${ENV_NAME} ${rdsName} Downloading finished OK. The size of ${temporaryfile} = ${temporaryfilesize} bytes. Start to add it into  ${commontemporary}"
+                            info "INFO: ${ENV_NAME} ${rdsName} Free disk space after downloading ${temporaryfile} df -h " : $(df -h)
+                            cat ${temporaryfile} >> ${commontemporary}
+                            echo >> ${commontemporary}
+                            rm -r ${temporaryfile}
+                            commontemporaryfilesize=$(stat -c%s "$commontemporary")
+                            info "INFO: ${ENV_NAME} ${rdsName} Size of collecting file is $commontemporary = $commontemporaryfilesize bytes."
+                            trycounter=10
+                        fi
+                    fi
+                done
+            done
+        fi
+        info "INFO: ${ENV_NAME} ${rdsName} Finished downloading ${rdsName} slowlogs by hours"
+        if [[ "${nextstep}" = "yes" && -f "${commontemporary}" ]] ; then
+            info "INFO: ${ENV_NAME} ${rdsName}. Starting to digest collected file ${commontemporary} and add result into anemometer database"
+            /usr/bin/pt-query-digest --user=$ANEMOMETER_MYSQL_USER --password=$ANEMOMETER_MYSQL_PASSWORD \
+                                    --review h=$ANEMOMETER_MYSQL_HOST,D=$ANEMOMETER_MYSQL_DB,t=global_query_review \
+                                    --history h=$ANEMOMETER_MYSQL_HOST,D=$ANEMOMETER_MYSQL_DB,t=global_query_review_history \
+                                    --no-report --limit=0% \
+                                    --filter=" \$event->{Bytes} = length(\$event->{arg}) and \$event->{hostname}=\"${rdsName}\"" \
+                                    ${commontemporary}
+            statuscode=$?
+            info "statuscode=${statuscode} of percona digest tool"
+            if [ ${statuscode} -gt 0 ] ; then
+                echo "CRITICAL: ${ENV_NAME} ${rdsName} slowlogs digest problem. the statuscode=${statuscode} "
+            else
+                info "INFO: ${ENV_NAME} ${rdsName} Digest of ${commontemporary} was successful"
+                rm -f "${commontemporary}"
             fi
-        done
+        fi
     done
 done